xorg-server-9999: fix kdrive build, and don't force opengl implem.
Kdrive needed all its possible servers added to IUSE_VIDEO_CARDS, which
was reduced to only released drivers ported to pciaccess. It was also
broken on amd64 because amd64 has vm86.h but the ebuild thought it
couldn't do vesa.

Re-enable and update patches from older versions for putting all kdrive
servers on a single line and building libxf86config PIC.

Remove the forcing to xorg-x11 by eselect-opengl because at this point,
it should build with anything, and this was causing some strange
failures during pkg_setup() because of the requirement for writability
to /.