xorg-server: Add 1.4.99.902.
Includes patches from
http://cgit.freedesktop.org/~dberkholz/xserver/log/?h=server-1.5-branch-gentoo
to port the swrast DRI driver to the 1.5 branch so we no longer require
mesa source at build time. That branch also includes a few more
HAL-related fixes to allow any xkb settings in the HAL FDI file.

It depends on some driver versions that aren't in the tree yet, but I'll
try to bump them soon. The more common ones are already around.

Update all X dependencies to the latest available versions, because
that's what upstream supports.

Xprint still doesn't work. Considering ripping out the xprint code,
because upstream source has ripped it out post-1.5 and intends for it to
be a separate package.