x-modular.eclass: Add EAPI support.
I decided to try something a little different because I had some ideas
for improving the existing EAPI patches I've seen going into other
eclasses. So here is my patch for x-modular.eclass. I tested it with
ebuilds using EAPIs 0, 1, and 2, and it appeared to work fine. It
already happened to have a function called src_configure, so that
doesn't appear in the patch.

One thing I see as an improvement is a lack of EAPI value checks
throughout the ebuild to avoid repetition between the function export
and the function call. Things just check whether a function was
exported, which is the only place where EAPI value checks happen.

Additionally, the fallback in case statements is "I don't know what to
do" and supported EAPIs are explicitly defined. This will make it
obvious when the eclass doesn't support a new EAPI instead of it
randomly failing after you try it.