x-modular.eclass: Reorganize for workaround w/ EXPORT_FUNCTIONS/inherit.
Move all code related to inherit up above the EXPORT_FUNCTIONS code,
instead of splitting EXPORT_FUNCTIONS into two separate places as the
previous patch did.

I talked to zmedico and he's fixing this in portage, but we're stuck w/
this order for a while unless we want to break people.