Various fixes for x11-drm, addition of x11-apps/drm-progs
Upstream x11-drm has moved dristat/drmstat out of the linux-core Makefile.
This prompted the removal of the last useful bits of the Makefile patch in our
patchball.

There are now no patches required to build these modules in a portage
environment, so the patchball is empty currently. As such, the logicused to
apply the patches has been changed. It now conditionally specifies a SRC_URI
only if PATCHVER is non-null, and will obviously not attempt to apply patches
if PATCHVER is null.

On a related note, the dristat and drmstat programs are no longer installed by
this package. They have been added as a separate package in this commit.

A minor fix to the MODULESD_... variables was made. It must strip the
.${KV_OBJ} from the module names, of course. The nv module has been removed
upstream, so the ebuild now offers nouveau via its own VIDEO_CARDS flag. The
radeonhd VIDEO_CARDS flag is now supported to allow building the radeon.ko
module without pulling in xf86-video-ati by VIDEO_CARDS=radeon dependencies.

A new package, x11-apps/drm-progs, has been added. This contains the dristat
and drmstat programs mentioned above, as well as a couple of modesetting tools.