[x11-apps/mesa-progs] Update the ebuild to match new upstream repo.
The mesa-demos tree has been separated from the main mesa repository.
I've updated the ebuild to use new repository location and tree.

As the upstream build system relies on unnecessary dependencies common
to all of the demos and we install only two small tools, I've decided to
avoid using it (and thus extremely raising the dependencies or hacking
the build system) and compile the tools using toolchain-funcs.