[ati-drivers] version bump and PowerXpress support added
Finally the PowerXpress support should be there.
Be aware i don't have the hardware to test on
so this work should be tested.
The switch can be done with amdcccle or aticonfig
tools. Needless to say only root can do that.
All the libs symlinks are removed since they
should not be needed with the custom switchlib{GL,glx}
scripts, instead eselect is used.

Many many thanks to gdbalbuzard [0] and chickpea [1]
for the help!

A null src_test was added since the package provides no test
functions.

[0] https://forums.gentoo.org/viewtopic-p-6748306.html?sid=789ed92ff252599668e53fbbcaf8b936#6748306
[1] https://bugs.sabayon.org/show_bug.cgi?id=2210